+++ benchmarks/runs/network_solver_benchmark_20260323_051249Z.json
@@ -1,0 +1,144 @@
+{
+  "run_utc": "2026-03-23T05:12:49.834352+00:00",
+  "repeats": 3,
+  "case_timeout_s": 300.0,
+  "solver_timeout_s": 120.0,
+  "cases": {
+    "grid_1x1_incomp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "Converged after fallback to hybr (original lm failed).",
+        "Converged after fallback to hybr (original lm failed).",
+        "Converged after fallback to hybr (original lm failed)."
+      ],
+      "mean_s": 0.17394270834241374,
+      "median_s": 0.17381412500981241,
+      "min_s": 0.17367345801903866,
+      "max_s": 0.1743405419983901,
+      "median_final_norm": 2.7725557891149334e-05,
+      "max_final_norm": 2.7725557891149334e-05
+    },
+    "grid_1x1_comp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 1.582243235995217,
+      "median_s": 1.582682082982501,
+      "min_s": 1.5782950000138953,
+      "max_s": 1.5857526249892544,
+      "median_final_norm": 0.0004941796373642008,
+      "max_final_norm": 0.0004941796373642008
+    },
+    "grid_2x2_incomp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 0.13785923601244576,
+      "median_s": 0.12574516600579955,
+      "min_s": 0.12340075001702644,
+      "max_s": 0.1644317920145113,
+      "median_final_norm": 6.682789239366259e-05,
+      "max_final_norm": 6.682789239366259e-05
+    },
+    "grid_2x2_comp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 5.748718236010366,
+      "median_s": 5.75382641601027,
+      "min_s": 5.705503833014518,
+      "max_s": 5.786824459006311,
+      "median_final_norm": 0.0008318515734930615,
+      "max_final_norm": 0.0008318515734930615
+    },
+    "grid_5x5_incomp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 0.7914337779899748,
+      "median_s": 0.8035397499916144,
+      "min_s": 0.7623991669970565,
+      "max_s": 0.8083624169812538,
+      "median_final_norm": 0.0003103075934512647,
+      "max_final_norm": 0.0003103075934512647
+    },
+    "grid_5x5_comp": {
+      "success_count": 0,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 0.0,
+      "messages": [
+        "Solver reported success but |F|=3.495e-03 exceeds residual tolerance (0.001).",
+        "Solver reported success but |F|=3.495e-03 exceeds residual tolerance (0.001).",
+        "Solver reported success but |F|=3.495e-03 exceeds residual tolerance (0.001)."
+      ],
+      "mean_s": 129.65593054199903,
+      "median_s": 129.64418899998418,
+      "min_s": 129.4016027089965,
+      "max_s": 129.92199991701636,
+      "median_final_norm": 0.0034951110669111476,
+      "max_final_norm": 0.0034951110669111476
+    },
+    "grid_10x10_incomp": {
+      "success_count": 0,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 0.0,
+      "messages": [
+        "Solver reported success but |F|=2.361e+00 exceeds residual tolerance (0.001).",
+        "Solver reported success but |F|=2.361e+00 exceeds residual tolerance (0.001).",
+        "Solver reported success but |F|=2.361e+00 exceeds residual tolerance (0.001)."
+      ],
+      "mean_s": 19.693914541344082,
+      "median_s": 19.653073541005142,
+      "min_s": 19.610125500010327,
+      "max_s": 19.818544583016774,
+      "median_final_norm": 2.360792851538123,
+      "max_final_norm": 2.360792851538123
+    },
+    "grid_10x10_comp": {
+      "success_count": 0,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 0.0,
+      "messages": [
+        "Solver reported success but |F|=1.594e+00 exceeds residual tolerance (0.001).",
+        "Solver reported success but |F|=1.594e+00 exceeds residual tolerance (0.001).",
+        "Solver reported success but |F|=1.594e+00 exceeds residual tolerance (0.001)."
+      ],
+      "mean_s": 199.8939927916605,
+      "median_s": 199.61004237501766,
+      "min_s": 199.50097920797998,
+      "max_s": 200.57095679198392,
+      "median_final_norm": 1.594237788549539,
+      "max_final_norm": 1.594237788549539
+    }
+  }
+}
