+++ benchmarks/runs/network_solver_benchmark_20260323_073319Z.json
@@ -1,0 +1,136 @@
+{
+  "run_utc": "2026-03-23T07:33:19.692773+00:00",
+  "repeats": 3,
+  "case_timeout_s": 300.0,
+  "solver_timeout_s": 120.0,
+  "cases": {
+    "grid_1x1_incomp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "Converged after fallback to hybr (original lm failed).",
+        "Converged after fallback to hybr (original lm failed).",
+        "Converged after fallback to hybr (original lm failed)."
+      ],
+      "mean_s": 0.16678426399206123,
+      "median_s": 0.16674612497445196,
+      "min_s": 0.16606979200150818,
+      "max_s": 0.16753687500022352,
+      "median_final_norm": 2.7725557891149334e-05,
+      "max_final_norm": 2.7725557891149334e-05
+    },
+    "grid_1x1_comp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 1.5200351526533875,
+      "median_s": 1.523481999989599,
+      "min_s": 1.5115300829638727,
+      "max_s": 1.5250933750066906,
+      "median_final_norm": 0.0004941796373642008,
+      "max_final_norm": 0.0004941796373642008
+    },
+    "grid_2x2_incomp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 0.1193237913151582,
+      "median_s": 0.11933187494287267,
+      "min_s": 0.11837670800741762,
+      "max_s": 0.12026279099518433,
+      "median_final_norm": 6.682789239366259e-05,
+      "max_final_norm": 6.682789239366259e-05
+    },
+    "grid_2x2_comp": {
+      "success_count": 3,
+      "attempts": 3,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100",
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 5.540646805350359,
+      "median_s": 5.511603291030042,
+      "min_s": 5.438314875005744,
+      "max_s": 5.672022250015289,
+      "median_final_norm": 0.0008318515734930615,
+      "max_final_norm": 0.0008318515734930615
+    },
+    "grid_5x5_incomp": {
+      "success_count": 1,
+      "attempts": 1,
+      "timeout_count": 0,
+      "success_rate": 1.0,
+      "messages": [
+        "The relative error between two consecutive iterates is at most 0.000100"
+      ],
+      "mean_s": 0.7408303749980405,
+      "median_s": 0.7408303749980405,
+      "min_s": 0.7408303749980405,
+      "max_s": 0.7408303749980405,
+      "median_final_norm": 0.0003103075934512647,
+      "max_final_norm": 0.0003103075934512647
+    },
+    "grid_5x5_comp": {
+      "success_count": 0,
+      "attempts": 1,
+      "timeout_count": 0,
+      "success_rate": 0.0,
+      "messages": [
+        "Solver reported success but |F|=3.495e-03 exceeds residual tolerance (0.001)."
+      ],
+      "mean_s": 132.2849300839589,
+      "median_s": 132.2849300839589,
+      "min_s": 132.2849300839589,
+      "max_s": 132.2849300839589,
+      "median_final_norm": 0.0034951110669111476,
+      "max_final_norm": 0.0034951110669111476
+    },
+    "grid_10x10_incomp": {
+      "success_count": 0,
+      "attempts": 1,
+      "timeout_count": 0,
+      "success_rate": 0.0,
+      "messages": [
+        "Solver reported success but |F|=2.361e+00 exceeds residual tolerance (0.001)."
+      ],
+      "mean_s": 19.57626583299134,
+      "median_s": 19.57626583299134,
+      "min_s": 19.57626583299134,
+      "max_s": 19.57626583299134,
+      "median_final_norm": 2.360792851538123,
+      "max_final_norm": 2.360792851538123
+    },
+    "grid_10x10_comp": {
+      "success_count": 0,
+      "attempts": 1,
+      "timeout_count": 0,
+      "success_rate": 0.0,
+      "messages": [
+        "Solver reported success but |F|=1.594e+00 exceeds residual tolerance (0.001)."
+      ],
+      "mean_s": 193.4446546669933,
+      "median_s": 193.4446546669933,
+      "min_s": 193.4446546669933,
+      "max_s": 193.4446546669933,
+      "median_final_norm": 1.594237788549539,
+      "max_final_norm": 1.594237788549539
+    }
+  }
+}
