+++ tests/test_thermo_transport.cpp
@@ -2721,21 +2721,21 @@ TEST_F(ThermoTransportTest, FrictionInvalidInputs) {
     EXPECT_THROW(friction_colebrook(10000.0, -0.001), std::invalid_argument);
 }
 
-// Test soft minimum constraint for Re <= 0
-TEST_F(ThermoTransportTest, FrictionSoftMinimum) {
+// Test smooth minimum constraint for Re <= 0
+TEST_F(ThermoTransportTest, FrictionSmoothMinimum) {
     double f_haaland_zero   = friction_haaland(0.0, 0.0);
     double f_haaland_neg    = friction_haaland(-1000.0, 0.0);
-    double f_haaland_min    = friction_haaland(100.0, 0.0);
-    double f_haaland_1000   = friction_haaland(1000.0, 0.0);
-    EXPECT_NEAR(f_haaland_zero, f_haaland_min, 1e-9);
-    EXPECT_NEAR(f_haaland_neg, f_haaland_1000, 1e-9);
+    double f_haaland_pos    = friction_haaland(1000.0, 0.0);
+    EXPECT_NEAR(f_haaland_neg, f_haaland_pos, 1e-9);
+    EXPECT_GT(f_haaland_zero, 0.0);
+    EXPECT_FALSE(std::isnan(f_haaland_zero));
 
     double f_colebrook_zero = friction_colebrook(0.0, 0.0);
     double f_colebrook_neg  = friction_colebrook(-1000.0, 0.0);
-    double f_colebrook_min  = friction_colebrook(100.0, 0.0);
-    double f_colebrook_1000 = friction_colebrook(1000.0, 0.0);
-    EXPECT_NEAR(f_colebrook_zero, f_colebrook_min, 1e-9);
-    EXPECT_NEAR(f_colebrook_neg, f_colebrook_1000, 1e-9);
+    double f_colebrook_pos  = friction_colebrook(1000.0, 0.0);
+    EXPECT_NEAR(f_colebrook_neg, f_colebrook_pos, 1e-9);
+    EXPECT_GT(f_colebrook_zero, 0.0);
+    EXPECT_FALSE(std::isnan(f_colebrook_zero));
 }
 
 // =============================================================================
